Move from the deprecated platform.linux_distribution() to the external package distro
platform.linux_distribution() has been deprecated some time ago, and the
only viable alternative seems to be this external package.
Add it as an optioanl dependency, as without it distribution detection
(i.e. `diffoscope --list-tools`) won't work.

Also add it as a Debian dependency, as it's a very lightway package, and
the usefulness it brings are great.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>